uporządkowanie funkcji pomocniczych
