+++ backend/routes/auth.py
@@ -6,39 +6,14 @@ from datetime import datetime, timedelta
 from dotenv import load_dotenv
 import os
 
-
+from utils.hashing import get_password_hash, verify_password
+from utils.tokenJWT import create_access_token
 from models import users as models
 from schemas import user as schemas  
 from database import get_db  
 
 
-# Wczytanie zmiennych z .env
-load_dotenv()
-
-SECRET_KEY = os.getenv("SECRET_KEY")
-ALGORITHM = os.getenv("ALGORITHM")
-ACCESS_TOKEN_EXPIRE_MINUTES = int(os.getenv("ACCESS_TOKEN_EXPIRE_MINUTES"))
-
-# Router
 router = APIRouter(tags=["Auth"])
-
-# Hashowanie haseł
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
-
-
-# Funkcje pomocnicze
-def get_password_hash(password: str) -> str:
-    return pwd_context.hash(password)
-
-def verify_password(plain_password: str, hashed_password: str) -> bool:
-    return pwd_context.verify(plain_password, hashed_password)
-
-def create_access_token(data: dict, expires_delta: timedelta = None):
-    to_encode = data.copy()
-    expire = datetime.utcnow() + (expires_delta or timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES))
-    to_encode.update({"exp": expire})
-    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
-
 
 # Rejestracja
 @router.post("/register", response_model=schemas.UserResponse)
